+++ app/Http/Controllers/admin/PrintersController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Printer;
 
@@ -11,6 +12,8 @@ class PrintersController extends Controller
   
     public function index()
     {
+        $this->authorize('view', new Printer);
+        
         $printers = Printer::all();
 
         return view('admin.printers.index', compact('printers'));
@@ -19,6 +22,8 @@ class PrintersController extends Controller
   
     public function create()
     {
+        $this->authorize('create', new Printer);
+
         return view('admin.printers.create');        
         
     }
@@ -26,6 +31,7 @@ class PrintersController extends Controller
   
     public function store(Request $request)
     {
+        $this->authorize('create', new Printer);
         
         //Validar el formulario
         $data = $request->validate([
@@ -51,18 +57,23 @@ class PrintersController extends Controller
  
     public function show(Printer $printer)
     {
+        $this->authorize('view', new Printer);
+
         return view('admin.printers.show', compact('printer'));
     }
 
     
     public function edit(Printer $printer)
     {
+        $this->authorize('update', $printer);
+
         return view('admin.printers.edit', compact('printer'));
         
     }
 
     public function update(Request $request, Printer $printer)
     {
+        $this->authorize('update', $printer);
         
         $printer->available = $request->available ? true : false;
         $printer->default = $request->default? true : false;
@@ -83,17 +94,24 @@ class PrintersController extends Controller
    
     public function destroy($idImpresora) //solo el admin hace esto
     {
+        $authUser = Auth::user(); // get current logged in user
         $printer = Printer::find($idImpresora); //busco al usuario a borrar
-
-       // $this->authorize('delete',$printer); // autorizo el delete, usando el policy
-
-        $printer->delete();
-
+       
+        if($authUser->can('delete',$printer)){ //si user autenticado puede borrar pd
+            $printer->delete();
+            $ok= true;
+            $mensaje='Impresora de tickets eliminada';
+        }else{
+            $ok= false;
+            $mensaje='No se puede eliminar la impresora de tickets';
+        }
         return response()->json(
             [
-            'ok' => true,
-            'mensaje' => 'La impresora ha sido borrado'
+            'ok' => $ok,
+            'mensaje' => $mensaje
             ]
         );
     }
+
+  
 }
